+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## TBD
+
+_Compatible with: [`smart-contract`](https://github.com/ndidplatform/smart-contract) v5.x.x_
+
+FEATURES:
+
+- `LOG_ONE_LINE` environment variable option for one-line pretty print when `LOG_PRETTY_PRINT` is set to `true`.
+
 ## 4.1.0 (June 28, 2020)
 
 _Compatible with: [`smart-contract`](https://github.com/ndidplatform/smart-contract) v5.x.x_